Use regular sbt dependencies for getting libgdx
instead of downloading a zip file with a sbt task, make use of regular
library dependency mechanism.

This means we depend on a specific (newer) version of libgdx. We
therefore need to move the asset folder.

We also move to sbt-native-packager for producing a runnable game. The
run script changes accordingly.